Removed all compiler warnings
Removed #windows included already warnings from glfw

+++ src/core/Camera.cpp
@@ -19,7 +19,7 @@ Camera::Camera(){
 	m_up = glm::vec3(0.0f, 1.0f, 0.0f);
 	m_right = glm::vec3(1.0f, 0.0f, 0.0f);
 
-	this->fov = TO_RAD(90.f);
+	this->fov = static_cast<float>(TO_RAD(90.f));
 	this->aspect_ratio = static_cast<float>(Config::_WINDOWWIDTH / Config::_WINDOWHEIGHT);
 	this->near_plane = 0.25f;
 	this->far_plane = 4096.f;
@@ -45,7 +45,7 @@ Camera::Camera(const char* name){
 	m_target = glm::vec3(0.0f, 0.0f, 1.0f);
 	m_up     = glm::vec3(0.0f, 1.0f, 0.0f);
 
-	this->fov = TO_RAD(90.f);
+	this->fov = static_cast<float>(TO_RAD(90.f));
 	this->aspect_ratio = static_cast<float>(Config::_WINDOWWIDTH / Config::_WINDOWHEIGHT);
 	this->near_plane = 0.25f;
 	this->far_plane = 4096.f;
